C: smarter multi-write system (splits it over multiple frames/loop passes)
2 write count systems:
2) MAX_WRITES: if write is not successful, how many times to keep trying
1) min_writes: how many successful writes should we send (likely not needed if [1] works correctly)

As of right now, writes are sort of slow so it makes sense to split it over multiple frames. Also gives a slight pause in writes to the motors. And even if writing becomes faster, then the loop runs faster and the next write comes faster so it shouldn't ever be a downside to split it up.